Add Initial Project Configration
This commit configures all dependencies and sets up a template that allows for connection to a database to create users and access said users.

The home screen includes a greeting and a hint towards accessing some sites that interact with the database. By adding a slash and a number to the URL, the app will look up the user with that ID in the database and display a greeting with that user's information.