Change "Simplified" methods to be the standard
This commit removes the "simplified" from some UserService/DAO methods and makes them the standard, while renaming the previous standard methods into having a "WithDetails" suffix. This should put the more efficient methods under the spotlight, forcing the programmers to use the less efficient versions only when necessary. Additionally, services for the other models have a findByName method that receives the backlog as a parameter.