Add separate backlog logic for logged in users
This commit changes some service methods so they receive both a backlog String and a User to decide what to do depending on whether the user is null or not. This allows the backlog stored in cookies to be used only when logged out. Currently, the same methods don't do anything if the user is logged in, and database access should be implemented next.