Remove empty lists from home page
This commit adds a condition to check whether a list has games in it before it displays it or not, so the home page isn't packed with empty lists. The only list that will always be shown will be Upcoming Games, since there should be at least one list in the home page and that one should be populated regardless of user activity on the site.