Remove ability to resubmit backlog change
This commit disables all add/remove buttons for a game while it is being added/removed to/from a user's backlog. The user may still add/remove many games at once before the page reloads, but once a game is selected, the action can't be undone by clicking the button again until the page reloads and the button shows the appropriate prompt to reverse that action.